+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-27T11:28:50.372Z")
+    created: Optional[datetime] = parse_datetime("2025-10-27T13:30:26.999Z")
     r"""The date and time the resource was created"""