+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8JPVD59WWNDMZ5MC8SBCKJ2"
+        "01K8JXT2NV4YPAR2NCT4VJRDDX"
     )
     r"""The unique identifier of the data source"""