+++ packages/orq-rc/RELEASES.md
@@ -3659,3 +3659,13 @@ Based on:
 - [python v3.14.0-rc.61] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.61] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.61 - packages/orq-rc
+
+## 2025-10-27 13:32:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.3 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.62] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.62] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.62 - packages/orq-rc